requirements.txt
streamlit
pandas
pyodbc
python-pptx
Pillow
requests
python-dotenv